fix new subroutine: "hwclock_modify"
After having this function for over one year (and finally pushing it yesterday),
I have to admit the previous code is a shame!
The 24bit part is 90% bug (typo in mask for negatives and missing carry) and
the rest is simply a bloated mess.
We must have survived this year out of dumb luck!?

Fixing and optimizing everything now...

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>