allow the LSB timer to be smaller than the MSB one
In order to save the two PINs (for PWM and T0) and the
electrical connection between them, it is possibe to prescale
timer1 by 256 and use some other 8-bit timer to follow
the prescaler synchronously.
The disadvantage is higher complexity accessing "hwclock_now"
(due to smaller LSB timer than MSB timer),
and calibrating the lower timer to the prescaler overflow.

Because it is not possibe to know or change the state of
prescalers, this approach needs some initial calibration.
Timer1 needs to be precisely watched in order to initialize the
LSB timer correctly.
For example read Timer1 lowbyte, read Timer1 lowbyte again
check if changed and set LSB timer value regarding cycles
consumed for check and set.

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>